requirements.txt
python-dotenv
pandas
numpy
pymongo[srv]>=4.0
certifi
scikit-learn
mlflow
pyyaml
dagshub
fastapi
uvicorn
jinja2
python-multipart